rfkill: Monitor WwanEnabled property changes properly
g-signal is not fired on property changes, so nm_signal is never called,
and all changes of NetworkManager's WwanEnabled property are missed.
That leads to an inability to disable the airplane mode on laptops with
WWAN after the airplane mode was once enabled by button.

Fix this by subscribing to g-properties-changed to actually listen to
the updates of WwanEnabled. It also makes the code simpler.

The bug was caused by a breaking change in NetworkManager [1] when it
deprecated the custom signal in favor of the standard PropertiesChanged
on org.freedesktop.DBus.Properties, and gnome-settings-daemon still used
the old one. The standard one is supported since NetworkManager 1.2.0,
as noted here [2].

[1]: https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/merge_requests/853
[2]: https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/commit/6fb917178aa19c61e909957f5146aa4565e0cb2f

Fixes: d74b6de142a2 ("rfkill: watch and set NetworkManager wwan-enabled")